Defective or not, pong is working. Lots of robustness missing in the language but its becoming feature complete for a POC.
Added rect math utilities and a byte array for storing data.

+++ src/RazerCore/window/SDLWindowImpl.cpp
@@ -71,7 +71,7 @@ void SDLWindowImpl::poll() {
         }
         if (e.type == SDL_KEYUP){
             m_keys[e.key.keysym.sym] = false;
-            RZLOG("key up %d\n", e.key.keysym.sym);
+            RZLOG("key up %d\n", (int)e.key.keysym.sym);
         }
         //If user clicks the mouse
         if (e.type == SDL_MOUSEBUTTONDOWN){
